Web Interface now shows icons and added support for user html templates
Images and stylesheets associated with a html template should now be placed in subfolders using the same name as the template to which they belong.

Added displaying of images in web view. Showing can be configured independantly from feed view.

You can now pas template=<template> as url argument in the feed link to use the imageset of the selected template as icons.

Replaced the default images with black on white versions.

added an option to show a replacement cover for when cover is missing for a single book.