Submit search now returns results that can be followed in the same way as typeahead results
From a multi library search, instead of each result link leading to the total list of books with a matching criteria,
A result link will now open the selected library page with the individual result counts for each category, allowing
you to zoom in on only the matches in the desired category,

If you start the search within a library, the behavior is exactly the same. You will remain on the category selection
page and the result count in each category will be shown under the category name just like on a multi-library page.

Some existing localization entries were renamed to prevent the need for new localization entries with same values.