Added First and Last navigation links to the footer.
Contrary to the Previous / Next Links they are not directly visible but will only show on pages 3 up to and including Max Page-2.
This was done since there is no sense in showing these links on pages where Previous or Next page has the same result.

added two entries for localization of title tooltip.